Edited Marshmallow Permission Request Flow
Solved the issue with denying the location permission. If user denys and the clicks on search devices, the permission request will be asked again.

Solved one major issue with searching of devices. The app didn't show the list of devices if another ble device is in the range. Searching was aborted because found device is not Neyya.